agent/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "accounting-agent"
version = "0.1.0"
description = "A minimal LangChain + LangGraph accounting agent demo."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "langchain>=1.2.15",
  "langchain-core>=1.2.30",
  "langchain-openai>=1.1.13",
  "langgraph>=1.1.6",
  "fastapi>=0.116.1",
  "httpx>=0.28.1",
  "uvicorn>=0.35.0",
  "python-multipart>=0.0.20",
  "python-dotenv>=1.0.1",
  "langsmith>=0.7.0",
  "langgraph-cli[inmem]>=0.4.23",
  "ipython>=8.39.0",
]

[project.scripts]
accounting-agent = "main:main"
accounting-agent-api = "account_agent.api.app:main"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
include = ["account_agent*"]

[dependency-groups]
dev = [
    "langgraph-cli>=0.4.23",
]
